refixed addressing modes and alu register descriptions, working on expanding fetch to support up to 8 bytes

+++ Emulator/include/debug.h
@@ -4,7 +4,7 @@
 #include "emulator/flags.h"
 
 void print_cmp_flags(CMP_Flags flags);
-void print_individual_bytes(u32 value);
+void print_individual_bytes(u64 value);
 
 #ifdef DEBUG
     #define d_printf(fmt, ...) printf(fmt, ##__VA_ARGS__)
