+++ Emulator/include/pipeline/fetch.h
@@ -4,6 +4,6 @@
 #include "my_stdtypes.h"
 #include "emulator/emulator.h"
 
-u32 fetch(Emulator *emu);
+u64 fetch(Emulator *emu);
 
 #endif
